double in testing

+++ pytorch/mnist_torch_testcase.py
@@ -140,7 +140,7 @@ def test(dataloader, model, loss_fn, device):
     test_loss, correct = 0, 0
     with torch.no_grad():
         for X, y in dataloader:
-            X, y = X.to(device), y.to(device)
+            X, y = X.to(device).double(), y.to(device)
             pred = model(torch.flatten(X, start_dim=1))
             test_loss += loss_fn(pred, y).item()
             correct += (pred.argmax(1) == y).type(torch.float).sum().item()
